Fix volume slider crash when decimal value received
The mini player volume slider has stepSize=1 but was receiving
unrounded float values like 98.666664 from device volume calculation,
causing IllegalStateException during view layout.

Added .toInt().toFloat() to round volume percentage same as the
syncSliderWithDeviceVolume() function already does.

+++ android/app/src/main/java/com/sendspindroid/MainActivity.kt
@@ -892,9 +892,10 @@ class MainActivity : AppCompatActivity() {
             )
 
             // Sync volume slider with device volume
+            // Round to nearest integer - slider has stepSize=1.0 and crashes on decimal values
             val maxVolume = audioManager.getStreamMaxVolume(AudioManager.STREAM_MUSIC)
             val currentVolume = audioManager.getStreamVolume(AudioManager.STREAM_MUSIC)
-            val volumePercent = if (maxVolume > 0) (currentVolume * 100f / maxVolume) else 50f
+            val volumePercent = if (maxVolume > 0) (currentVolume * 100f / maxVolume).toInt().toFloat() else 50f
             miniPlayer.miniPlayerVolumeSlider?.value = volumePercent
 
             // Load album art if available
